+++ src/pdm_conda/project.py
@@ -44,8 +44,9 @@ class CondaProject(Project):
         self.core.resolver_class = CondaResolver
         self.locked_repository_class = LockedCondaRepository
         self.environment_class = CondaEnvironment
-        self.virtual_packages: set[CondaRequirement] = set()
+        self.virtual_packages: set[CondaRequirement] | None = None
         self.platform: str | None = None
+        self.default_channels: list[str] | None = None
         self._conda_mapping: dict[str, str] = dict()
         self._pypi_mapping: dict[str, str] = dict()
         self.conda_config = PluginConfig.load_config(self)