+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.8.1]
+
+### Changed
+
+* Use Conda channels priority in package sorting.
+* `PyProject` now contains info from default_channels, virtual_packages and platform.
+
 ## [0.8.0] - 09/04/2023
 
 ### Added